+++ src/Symfony/Bundle/FrameworkBundle/Command/EventDispatcherDebugCommand.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Command;
 
+use Psr\Container\ContainerInterface;
 use Symfony\Bundle\FrameworkBundle\Console\Helper\DescriptorHelper;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -29,14 +30,16 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
  */
 class EventDispatcherDebugCommand extends Command
 {
-    protected static $defaultName = 'debug:event-dispatcher';
-    private $dispatcher;
+    private const DEFAULT_DISPATCHER = 'event_dispatcher';
 
-    public function __construct(EventDispatcherInterface $dispatcher)
+    protected static $defaultName = 'debug:event-dispatcher';
+    private $dispatchers;
+
+    public function __construct(ContainerInterface $dispatchers)
     {
         parent::__construct();
 
-        $this->dispatcher = $dispatcher;
+        $this->dispatchers = $dispatchers;
     }
 
     /**
@@ -47,6 +50,7 @@ class EventDispatcherDebugCommand extends Command
         $this
             ->setDefinition([
                 new InputArgument('event', InputArgument::OPTIONAL, 'An event name or a part of the event name'),
+                new InputOption('dispatcher', null, InputOption::VALUE_REQUIRED, 'To view events of a specific event dispatcher', self::DEFAULT_DISPATCHER),
                 new InputOption('format', null, InputOption::VALUE_REQUIRED, 'The output format  (txt, xml, json, or md)', 'txt'),
                 new InputOption('raw', null, InputOption::VALUE_NONE, 'To output raw description'),
             ])
@@ -74,12 +78,21 @@ EOF
         $io = new SymfonyStyle($input, $output);
 
         $options = [];
+        $dispatcherServiceName = $input->getOption('dispatcher');
+        if (!$this->dispatchers->has($dispatcherServiceName)) {
+            $io->getErrorStyle()->error(sprintf('Event dispatcher "%s" is not available.', $dispatcherServiceName));
+
+            return 1;
+        }
+
+        $dispatcher = $this->dispatchers->get($dispatcherServiceName);
+
         if ($event = $input->getArgument('event')) {
-            if ($this->dispatcher->hasListeners($event)) {
+            if ($dispatcher->hasListeners($event)) {
                 $options = ['event' => $event];
             } else {
                 // if there is no direct match, try find partial matches
-                $events = $this->searchForEvent($this->dispatcher, $event);
+                $events = $this->searchForEvent($dispatcher, $event);
                 if (0 === \count($events)) {
                     $io->getErrorStyle()->warning(sprintf('The event "%s" does not have any registered listeners.', $event));
 
@@ -93,10 +106,15 @@ EOF
         }
 
         $helper = new DescriptorHelper();
+
+        if (self::DEFAULT_DISPATCHER !== $dispatcherServiceName) {
+            $options['dispatcher_service_name'] = $dispatcherServiceName;
+        }
+
         $options['format'] = $input->getOption('format');
         $options['raw_text'] = $input->getOption('raw');
         $options['output'] = $io;
-        $helper->describe($io, $this->dispatcher, $options);
+        $helper->describe($io, $dispatcher, $options);
 
         return 0;
     }