+++ src/Symfony/Bundle/FrameworkBundle/Console/Descriptor/MarkdownDescriptor.php
@@ -287,8 +287,14 @@ class MarkdownDescriptor extends Descriptor
     protected function describeEventDispatcherListeners(EventDispatcherInterface $eventDispatcher, array $options = [])
     {
         $event = \array_key_exists('event', $options) ? $options['event'] : null;
+        $dispatcherServiceName = $options['dispatcher_service_name'] ?? null;
 
         $title = 'Registered listeners';
+
+        if (null !== $dispatcherServiceName) {
+            $title .= sprintf(' of event dispatcher "%s"', $dispatcherServiceName);
+        }
+
         if (null !== $event) {
             $title .= sprintf(' for event `%s` ordered by descending priority', $event);
             $registeredListeners = $eventDispatcher->getListeners($event);