+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 -----
 
  * Added support for configuring PHP error level to log levels
+* Added the `dispatcher` option to `debug:event-dispatcher`
+* Added the `event_dispatcher.dispatcher` tag
 
 5.2.0
 -----